Update bloc and meta dependencies to newer versions
This commit updates the `bloc` and `meta` dependencies to their newer
versions. The `bloc` dependency is updated from `2.11.0` to `8.1.2`, and
the `meta` dependency is updated from `1.9.1` to `1.10.0`. These updates
bring in bug fixes, performance improvements, and new features provided
by the updated versions of these packages.

No issues are associated with this commit.